Move Serialisable to axioms, rename transparency to repr_transparency
Per review: `Category` is the first class of `discopy.abc` again, and the
serialisation interface lives at the end of `discopy.axioms`, beside the
machinery that states its laws. `abc` keeps re-exporting it, as it does
`Equation` and `NamedGeneric`, so `discopy.abc.Serialisable` still
resolves; nothing in `abc` referred to it but the import.

The representation roundtrip is `repr_transparency`, leaving room for
the `str` one: STYLE.md asks that `eval(str(x)) == x` under the obvious
naming convention, but a `str` prints bare names, so the environment it
reads back in has to be built from the term rather than from a module.
That is #764, which the docstring links.
